# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: A²ES - AI for Energy & Sustainability API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /equipment:
    get:
      operationId: listEquipment
      tags: [equipment]
      summary: List all equipment
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: efficiency
          in: query
          schema:
            type: string
            enum: [inefficient, average, optimized]
        - name: sector
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of equipment
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Equipment"
    post:
      operationId: createEquipment
      tags: [equipment]
      summary: Create new equipment entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEquipmentBody"
      responses:
        "201":
          description: Equipment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
  /equipment/{id}:
    get:
      operationId: getEquipment
      tags: [equipment]
      summary: Get equipment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Equipment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
        "404":
          description: Not found
    put:
      operationId: updateEquipment
      tags: [equipment]
      summary: Update equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEquipmentBody"
      responses:
        "200":
          description: Equipment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    delete:
      operationId: deleteEquipment
      tags: [equipment]
      summary: Delete equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /chat/sessions:
    get:
      operationId: listChatSessions
      tags: [chat]
      summary: List chat sessions
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSession"
    post:
      operationId: createChatSession
      tags: [chat]
      summary: Create a new chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatSessionBody"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
  /chat/sessions/{id}/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get messages for a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to AI expert (streaming SSE)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatMessageBody"
      responses:
        "200":
          description: SSE stream of response chunks
          content:
            text/event-stream:
              schema:
                type: string
...
components:
  schemas:
...
    Equipment:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        brand:
          type: string
        model:
          type: string
        yearManufactured:
          type: integer
          nullable: true
        voltage:
          type: string
          nullable: true
        current:
          type: string
          nullable: true
        power:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        size:
          type: string
          nullable: true
        hoursPerDay:
          type: number
          nullable: true
        daysPerWeek:
          type: number
          nullable: true
        sector:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        operatorNotes:
          type: string
          nullable: true
        apparentCondition:
          type: string
          nullable: true
        noiseObserved:
          type: boolean
          nullable: true
        heatingObserved:
          type: boolean
          nullable: true
        vibrationObserved:
          type: boolean
          nullable: true
        imageUrl:
          type: string
          nullable: true
        efficiencyLevel:
          type: string
          nullable: true
          enum: [inefficient, average, optimized]
        estimatedPowerKw:
          type: number
          nullable: true
        estimatedCo2MonthlyKg:
          type: number
          nullable: true
        estimatedConsumptionMonthlyKwh:
          type: number
          nullable: true
        retrofitStatus:
          type: string
          nullable: true
          enum: [pending, in_progress, completed, not_needed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - brand
        - model
        - createdAt
        - updatedAt
    CreateEquipmentBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        brand:
          type: string
        model:
          type: string
        yearManufactured:
          type: integer
          nullable: true
        voltage:
          type: string
          nullable: true
        current:
          type: string
          nullable: true
        power:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        size:
          type: string
          nullable: true
        hoursPerDay:
          type: number
          nullable: true
        daysPerWeek:
          type: number
          nullable: true
        sector:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        operatorNotes:
          type: string
          nullable: true
        apparentCondition:
          type: string
          nullable: true
        noiseObserved:
          type: boolean
          nullable: true
        heatingObserved:
          type: boolean
          nullable: true
        vibrationObserved:
          type: boolean
          nullable: true
        imageUrl:
          type: string
          nullable: true
      required:
        - name
        - type
        - brand
        - model
    UpdateEquipmentBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        brand:
          type: string
        model:
          type: string
        yearManufactured:
          type: integer
          nullable: true
        voltage:
          type: string
          nullable: true
        current:
          type: string
          nullable: true
        power:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        size:
          type: string
          nullable: true
        hoursPerDay:
          type: number
          nullable: true
        daysPerWeek:
          type: number
          nullable: true
        sector:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        operatorNotes:
          type: string
          nullable: true
        apparentCondition:
          type: string
          nullable: true
        noiseObserved:
          type: boolean
          nullable: true
        heatingObserved:
          type: boolean
          nullable: true
        vibrationObserved:
          type: boolean
          nullable: true
        imageUrl:
          type: string
          nullable: true
        efficiencyLevel:
          type: string
          nullable: true
          enum: [inefficient, average, optimized]
        retrofitStatus:
          type: string
          nullable: true
          enum: [pending, in_progress, completed, not_needed]
...
    ChatSession:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        equipmentId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    CreateChatSessionBody:
      type: object
      properties:
        title:
          type: string
        equipmentId:
          type: integer
          nullable: true
      required:
        - title
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - role
        - content
        - createdAt
    SendChatMessageBody:
      type: object
      properties:
        content:
          type: string
        equipmentContext:
          type: string
          nullable: true
      required:
        - content
...