# lib/api-zod/src/generated/types/createChatSessionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * A²ES - AI for Energy & Sustainability API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateChatSessionBody {
  title: string;
  equipmentId?: number | null;
}