# lib/api-zod/src/generated/types/updateEquipmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * A²ES - AI for Energy & Sustainability API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateEquipmentBodyEfficiencyLevel } from "./updateEquipmentBodyEfficiencyLevel";
import type { UpdateEquipmentBodyRetrofitStatus } from "./updateEquipmentBodyRetrofitStatus";

export interface UpdateEquipmentBody {
  name?: string;
  type?: string;
  brand?: string;
  model?: string;
  yearManufactured?: number | null;
  voltage?: string | null;
  current?: string | null;
  power?: string | null;
  color?: string | null;
  size?: string | null;
  hoursPerDay?: number | null;
  daysPerWeek?: number | null;
  sector?: string | null;
  location?: string | null;
  operatorNotes?: string | null;
  apparentCondition?: string | null;
  noiseObserved?: boolean | null;
  heatingObserved?: boolean | null;
  vibrationObserved?: boolean | null;
  imageUrl?: string | null;
  efficiencyLevel?: UpdateEquipmentBodyEfficiencyLevel;
  retrofitStatus?: UpdateEquipmentBodyRetrofitStatus;
}